iban instead of email for transactions

+++ src/main/java/com/hugos/BanKING/services/ResourceProtectionService.java
@@ -1,10 +1,16 @@
 package com.hugos.BanKING.services;
 
+import com.hugos.BanKING.entities.BankAccount;
 import com.hugos.BanKING.enums.Role;
+import com.hugos.BanKING.repositories.BankAccountRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
 import javax.servlet.http.HttpServletRequest;
+import java.util.Optional;
 
 // This class is here to make sure any user trying to access a resource actually has access to that resource
 // This verification process includes: JWT verification and Role validation
@@ -17,6 +23,7 @@ public class ResourceProtectionService {
     private final AppUserService appUserService;
     private final RequestService requestService;
     private final TransactionService transactionService;
+    private final BankAccountRepository bankAccountRepository;
 
     public ResponseEntity<?> getAppUser(HttpServletRequest request, String email) {
         requestService.authorizeRequest(request, Role.USER, email);
@@ -40,22 +47,42 @@ public class ResourceProtectionService {
         return appUserService.authenticateAppUser(request);
     }
 
-    public ResponseEntity<?> createTransaction(HttpServletRequest request, String email, String type) {
+    public ResponseEntity<?> createTransaction(HttpServletRequest request, String iban, String type) {
+        Optional<BankAccount> bankAccountOptional = bankAccountRepository.findByIban(iban);
+        if (bankAccountOptional.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Bank account with iban not found");
+        }
+        String email = bankAccountOptional.get().getAppUser().getEmail();
         requestService.authorizeRequest(request, Role.USER, email);
         return transactionService.createTransaction(request, email, type);
     }
 
-    public ResponseEntity<?> getTransactions(HttpServletRequest request, String email, String sortBy) {
+    public ResponseEntity<?> getTransactions(HttpServletRequest request, String iban, String sortBy) {
+        Optional<BankAccount> bankAccountOptional = bankAccountRepository.findByIban(iban);
+        if (bankAccountOptional.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Bank account with iban not found");
+        }
+        String email = bankAccountOptional.get().getAppUser().getEmail();
         requestService.authorizeRequest(request, Role.USER, email);
         return transactionService.getTransactions(email, sortBy);
     }
 
-    public ResponseEntity<?> updateTransaction(HttpServletRequest request, String email, Long id) {
+    public ResponseEntity<?> updateTransaction(HttpServletRequest request, String iban, Long id) {
+        Optional<BankAccount> bankAccountOptional = bankAccountRepository.findByIban(iban);
+        if (bankAccountOptional.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Bank account with iban not found");
+        }
+        String email = bankAccountOptional.get().getAppUser().getEmail();
         requestService.authorizeRequest(request, Role.ADMIN, email);
         return transactionService.updateTransaction(request, email, id);
     }
 
-    public ResponseEntity<?> deleteTransactions(HttpServletRequest request, String email) {
+    public ResponseEntity<?> deleteTransactions(HttpServletRequest request, String iban) {
+        Optional<BankAccount> bankAccountOptional = bankAccountRepository.findByIban(iban);
+        if (bankAccountOptional.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Bank account with iban not found");
+        }
+        String email = bankAccountOptional.get().getAppUser().getEmail();
         requestService.authorizeRequest(request, Role.USER, email);
         return transactionService.deleteTransactions(email);
     }
